Parcours reset

+++ back/parcours_imi/admin/user_parcours_reset.py
@@ -4,12 +4,13 @@ from django.contrib import messages
 from django.shortcuts import render, redirect
 
 from parcours_imi.models import UserParcours
+from parcours_imi.tasks import send_user_parcours_reset_email
 
 
 logger = logging.getLogger(__name__)
 
 
-def user_parcours_unlock_courses_view(request, object_id):
+def user_parcours_reset_view(request, object_id):
     try:
         user_parcours = UserParcours.objects.get(pk=object_id)
     except UserParcours.DoesNotExist:
@@ -21,10 +22,12 @@ def user_parcours_unlock_courses_view(request, object_id):
     try:
         course_choice = user_parcours.course_choice
     except UserParcours.course_choice.RelatedObjectDoesNotExist:
-        messages.warning(request, 'L\'étudiant n\'a pas encore choisi ses cours.')
-        return redirect('admin:parcours_imi_userparcours_change', object_id=object_id)
+        course_choice = None
 
-    course_choice.delete()
+    if course_choice:
+        course_choice.delete()
 
-    messages.success(request, 'Cours dévérouillés avec succès.')
+    send_user_parcours_reset_email.delay(object_id)
+
+    messages.success(request, 'Parcours réinitialisé avec succès.')
     return redirect('admin:parcours_imi_userparcours_change', object_id=object_id)
